Bugfix on houses offset calculation.
We were only using integer when computing a relative position
between 0 and 1 for a house on a road chunk. This was leading
to a misplacement, houses were always at the end (or start) of the chunk.

+++ src/34_road/RoadChunk.cpp
@@ -274,8 +274,8 @@ namespace synthese
 			return
 				startOffset + (endOffset - startOffset) * (
 					(bounds->first < bounds->second) ?
-					((houseNumber - bounds->first) / (bounds->second - bounds->first)) :
-					((bounds->first - houseNumber) / (bounds->first - bounds->second))
+					((double)(houseNumber - bounds->first) / (double)(bounds->second - bounds->first)) :
+					((double)(bounds->first - houseNumber) / (double)(bounds->first - bounds->second))
 				)
 			;
 		}
